Fix formatting and include paths for linters (#157)
* Fix uncrustify
* Add deprecation notes
* Try to remove filesystem helper
* Change method return type to size_t
* Fix MSVC warning for getenv
* Rename .h files to .hpp + formatting fixes
- fix include guards to cpplint style
- fix order of includes to match ros guide
* Add back deprecated .h headers
* Change all getNumSubscriber/Publisher signatures to size_t
* Fix leaky windows workaround for getenv
* Fixes to ament dependencies.
* Use rcpputils::get_env_var rather than custom
* Address reviewer feedback
* Re-add rcpputils for get_env_var
* Change to new style remapping
* Move dllimport/export locations

+++ image_transport/src/camera_publisher.cpp
@@ -32,13 +32,14 @@
 *  POSSIBILITY OF SUCH DAMAGE.
 *********************************************************************/
 
-#include "image_transport/camera_publisher.h"
-#include "image_transport/image_transport.h"
-#include "image_transport/camera_common.h"
+#include "image_transport/camera_publisher.hpp"
 
 #include <rclcpp/expand_topic_or_service_name.hpp>
 #include <rclcpp/logging.hpp>
 #include <rclcpp/node.hpp>
+
+#include "image_transport/camera_common.hpp"
+#include "image_transport/image_transport.hpp"
 
 namespace image_transport
 {
@@ -95,7 +96,7 @@ CameraPublisher::CameraPublisher(
   impl_->info_pub_ = node->create_publisher<sensor_msgs::msg::CameraInfo>(info_topic, qos);
 }
 
-uint32_t CameraPublisher::getNumSubscribers() const
+size_t CameraPublisher::getNumSubscribers() const
 {
   //TODO(ros2) add support when rcl/rmw support it.
   //if (impl_ && impl_->isValid())
